feat(frontend): unified course search in planner using timetable search modal
- Extract CourseSearch into shared common component with toggleMode prop
- Timetable course-search becomes thin wrapper around shared component
- New PlannerCourseSearch: search modal + inline detail form for grade/category
- Replace AddCourseForm in SemesterPanel with PlannerCourseSearch
- Initialize ApiProvider globally in AppShell for cross-page availability
